+++ dacite/src/components/routing/routing-helpers.ts
@@ -1,28 +1,14 @@
+import { NextRouter } from "next/router";
 import pg from "../../db";
-import { QueryI } from "..";
+import { QueryI, CrumbsI } from "./base-page";
 
-async function getIdHierarchy(query: QueryI) {
-  // could be col_id,section_id, unit_id
-  const { project_id, col_id, section_id, unit_id, col_group_id } = query;
-
-  // Can you use simpler null checks?
-  // e.g., if (col_id != null)  could work well.
-  // Also this seems like a weird way to do it... this function
-  // seems like a choke point that just passes things off to other functions.
-  if (typeof col_id !== "undefined") {
-    return await fetchIdsFromColId(col_id);
-  } else if (typeof section_id !== "undefined") {
-    return await fetchIdsFromSectionId(section_id);
-  } else if (typeof unit_id !== "undefined") {
-    return await fetchIdsFromUnitId(unit_id);
-  } else if (typeof col_group_id !== "undefined") {
-    return await fetchIdsFromColGroup(col_group_id);
-  }
-
-  return query;
+export interface IdsFromColGroup {
+  project_id?: number;
 }
 
-async function fetchIdsFromColGroup(col_group_id: number) {
+async function fetchIdsFromColGroup(
+  col_group_id: number
+): Promise<IdsFromColGroup> {
   const { data, error } = await pg
     .from("col_groups")
     .select("project_id")
@@ -72,4 +58,67 @@ async function fetchIdsFromColId(col_id: number) {
   return {};
 }
 
-export { getIdHierarchy };
+interface BreadCrumbsHookI {
+  query: QueryI;
+  router: NextRouter;
+}
+function useBreadCrumbs(props: BreadCrumbsHookI) {
+  const { query, router } = props;
+
+  const filterCrumbs = (obj: CrumbsI): boolean => {
+    if (obj.text == "Projects") {
+      return true;
+    }
+
+    if (!(obj.predicate in query)) return false;
+    return true;
+  };
+
+  const breadCrumbs: CrumbsI[] = [
+    {
+      text: "Projects",
+      onClick: async () => {
+        router.push("/");
+      },
+      predicate: "",
+    },
+    {
+      text: "Column Groups",
+      onClick: async () => {
+        router.push(`/column-groups/${query.project_id}`);
+      },
+      predicate: "project_id",
+    },
+    {
+      text: "Column",
+      onClick: async () => {
+        router.push(`/column/${query.col_id}`);
+      },
+      predicate: "col_id",
+    },
+    {
+      text: "Section",
+      onClick: async () => {
+        router.push(`/section/${query.section_id}`);
+      },
+      predicate: "section_id",
+    },
+    {
+      text: "Unit",
+      onClick: async () => {
+        router.push(`/unit/${query.unit_id}/edit`);
+      },
+      predicate: "unit_id",
+    },
+  ].filter(filterCrumbs);
+
+  return breadCrumbs;
+}
+
+export {
+  fetchIdsFromColId,
+  fetchIdsFromSectionId,
+  fetchIdsFromUnitId,
+  fetchIdsFromColGroup,
+  useBreadCrumbs,
+};
